feat: adiciona testes pytest e corrige serialização np.int64

+++ main.py
@@ -29,12 +29,12 @@ def stats():
     resultado = {}
     for coluna in numericas.columns:
         resultado[coluna] = {
-            "media": round(numericas[coluna].mean(), 2),
-            "mediana": round(numericas[coluna].median(), 2),
-            "desvio_padrao": round(numericas[coluna].std(), 2),
-            "minimo": round(numericas[coluna].min(), 2),
-            "maximo": round(numericas[coluna].max(), 2)
-        }
+            "media": float(round(numericas[coluna].mean(), 2)),
+            "mediana": float(round(numericas[coluna].median(), 2)),
+            "desvio_padrao": float(round(numericas[coluna].std(), 2)),
+            "minimo": float(round(numericas[coluna].min(), 2)),
+            "maximo": float(round(numericas[coluna].max(), 2))
+}
 
     return jsonify(resultado)
